chore: initial refactor
