+++ v3/src/components/graph/components/graph-attribute-label.tsx
@@ -5,87 +5,17 @@ import { t } from "../../../utilities/translation/translate"
 import { useGraphDataConfigurationContext } from "../hooks/use-graph-data-configuration-context"
 import { useGraphContentModelContext } from "../hooks/use-graph-content-model-context"
 import { useGraphLayoutContext } from "../hooks/use-graph-layout-context"
+import { useTileSelectionContext } from "../../../hooks/use-tile-selection-context"
 import { AttributeType } from "../../../models/data/attribute-types"
 import { IDataSet } from "../../../models/data/data-set"
 import { GraphPlace, isVertical } from "../../axis-graph-shared"
-import { labelPadding } from "../../axis/axis-types"
+import { labelMargin } from "../../axis/axis-types"
+import { getStringBounds, renderLabelBackground } from "../../axis/axis-utils"
 import { AttributeLabel } from "../../data-display/components/attribute-label"
 import { graphPlaceToAttrRole } from "../../data-display/data-display-types"
-import { useTileSelectionContext } from "../../../hooks/use-tile-selection-context"
-import { getStringBounds } from "../../axis/axis-utils"
 import { ClickableAxisLabel } from "./clickable-axis-label"
 
 import vars from "../../vars.scss"
-
-interface IRenderLabelBackgroundOptions {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  gSelection: any  // D3 Selection<SVGGElement | null, ...> — typed as any to avoid variance issues
-  textSelector: string
-  transform?: string
-  visibility?: string
-}
-
-/**
- * Renders a background rect and dropdown arrow behind/after an axis or legend label text element.
- * The rect provides hover/focus/selected visual states via CSS classes on the parent <g>.
- * The arrow uses the same path as dropdown-arrow-icon.svg (10x5 triangle in a 24x24 viewBox).
- */
-function renderLabelBackground(
-  { gSelection, textSelector, transform = '', visibility }: IRenderLabelBackgroundOptions
-) {
-  const textNode = gSelection.select(textSelector).node() as SVGTextElement | null
-  const textBBox = textNode?.getBBox()
-  if (!textBBox) return
-
-  const paddingX = 8
-  const paddingY = 2
-  const arrowWidth = 24  // icon container size matching dropdown-arrow-icon.svg viewBox
-
-  // Background rect: left padding + text + arrow (arrow's 24x24 container provides right padding)
-  const rectWidth = textBBox.width + paddingX + arrowWidth
-  const rectHeight = textBBox.height + paddingY * 2
-  const rectX = textBBox.x - paddingX
-  const rectY = textBBox.y - paddingY
-
-  const rectSelection = gSelection.selectAll('rect.attribute-label-bg')
-    .data([1])
-    .join(
-      (enter: any) => enter.append('rect').attr('class', 'attribute-label-bg'),
-      (update: any) => update
-    )
-    .attr('x', rectX)
-    .attr('y', rectY)
-    .attr('width', rectWidth)
-    .attr('height', rectHeight)
-    .attr('rx', 4)
-  if (transform) rectSelection.attr('transform', transform)
-  if (visibility) rectSelection.style('visibility', visibility)
-  rectSelection.lower()  // Ensure rect renders behind text
-
-  // Dropdown arrow: nested <svg> using the same path as dropdown-arrow-icon.svg
-  // Place directly after text — the icon's built-in padding provides sufficient spacing
-  const arrowX = textBBox.x + textBBox.width
-  const arrowY = textBBox.y + (textBBox.height - arrowWidth) / 2
-
-  const arrowSelection = gSelection.selectAll('svg.attribute-label-arrow')
-    .data([1])
-    .join(
-      (enter: any) => {
-        const arrow = enter.append('svg')
-          .attr('class', 'attribute-label-arrow')
-          .attr('viewBox', '0 0 24 24')
-          .attr('width', arrowWidth)
-          .attr('height', arrowWidth)
-        arrow.append('path').attr('d', 'm12 15-5-5h10z')
-        return arrow
-      },
-      (update: any) => update
-    )
-    .attr('x', arrowX)
-    .attr('y', arrowY)
-  if (transform) arrowSelection.attr('transform', transform)
-  if (visibility) arrowSelection.style('visibility', visibility)
-}
 
 interface IAttributeLabelProps {
   place: GraphPlace
@@ -285,8 +215,8 @@ export const GraphAttributeLabel =
         labelBounds = getStringBounds(label, labelFont),
         labelTransform = `translate(${bounds.left}, ${bounds.top})`,
         // With dominant-baseline: central, tY sets the visual center of the text.
-        // Place center at labelPadding + labelBounds.height/2 from the outer edge.
-        labelCenter = labelPadding + labelBounds.height / 2,
+        // Place center at labelMargin + labelBounds.height/2 from the outer edge.
+        labelCenter = labelMargin + labelBounds.height / 2,
         tX = place === 'left' ? labelCenter
           : place === 'legend' ? bounds.left
             : ['rightNumeric', 'rightCat'].includes(place) ? bounds.width - labelCenter
